Replace motd.*, rejigger src/sys/conf/*.sh, and fix Makefile.bootcd.
I changed the version scheme to pretty much just
reflect the git branch it came from. This may be
changed to be a variation of the git shorthash, but for
now, this works. I also replaced the NetBSD motd files
with Lambda motds, and edited src/etc/Makefile to
accomodate this. Finally, the version scheme broke
building boot CDs because of my imprudent insistence on
using unicode in the version scheme, so I changed this
to not use the version in the volume ID. This is
probably not wise.